feat: restructuring banner files
After adding the reject cookies extended functionality, the original cookie-banner file has become difficult to read. This commit extracts the code into separate files, including cookie-functions helper file and files that contain constants
