+++ engine/app/assets/javascript/citizens_advice_cookie_preferences/application.js
@@ -1,2 +1,2 @@
-(()=>{function u(){window.dataLayer.push({event:"acceptAllCookies"})}var t={cookieBanner:".js-cookie-banner",acceptBtn:"#js-cookie-banner__button-accept",rejectBtn:"#js-cookie-banner__button-reject",cookieSelectionContainer:".js-cookie-banner__selection-container",confirmationMessageContainer:".js-cookie-banner__confirmation-message",confirmationMessageAccept:".js-cookie-banner__confirmation-accept",confirmationMessageReject:".js-cookie-banner__confirmation-reject",hideBannerBtn:"#js-cookie-banner__button-hide"},a={_smart_meter_tool_session:"essential",_demo_session:"essential",_content_platform_forms_session:"essential",_energy_apps_session:"essential",cookie_preference:"essential",cookie_preference_set:"essential",_dd_s:"essential",_forms_key:"essential",_grecaptcha:"essential",activeChat:"essential","amazon-connect-*":"essential","amazon-connect-session-*":"essential","cobrowse-storage_expiration-22834971":"essential",CustomizationObject:"essential",cwr_s:"essential",cwr_u:"essential","lpLastVisit-*":"essential",lpPmCalleeDfs:"essential","LPSID-*":"essential",lpTabId:"essential",LPVID:"essential",persistedChatSession:"essential","X-Source":"essential",_ga:"analytics","_ga_*":"analytics",ar_debug:"analytics",ethnio_displayed:"analytics",geo:"analytics","mf_*":"analytics",mf_initialDomQueue:"analytics",mf_transmitQueue:"analytics",mf_user:"analytics",mouseflow:"analytics","ca_ab_*":"analytics",__TAG_ASSISTANT:"analytics","some_*":"essential"},p={essential:!0,analytics:!0,video_players:!1},y={essential:!0,analytics:!1,video_players:!1},m=document.location.hostname==="localhost"?"localhost":"citizensadvice.org.uk";function c(e,n,i){let o=new Date;o.setTime(o.getTime()+i*24*60*60*1e3);let r="expires="+o.toUTCString();document.cookie=e+"="+n+";"+r+`;path=/;domain=${m}`}function f(e){let n=`${e}=`,i=document.cookie.split(";");for(let o=0,r=i.length;o<r;o++){let s=i[o];for(;s.charAt(0)===" ";)s=s.substring(1,s.length);if(s.indexOf(n)===0)return decodeURIComponent(s.substring(n.length))}return null}var C=()=>{c("cookie_preference",encodeURIComponent(JSON.stringify({essential:!0,analytics:!0,video_players:!0})),365),c("some_test_cookie",encodeURIComponent(JSON.stringify({something:!0})),365);var e=document.getElementsByClassName("js-cookie-banner")[0].getAttribute("data-cookie-current-version");c("cookie_preference_set",e,365),k(),document.querySelector(t.confirmationMessageAccept).hidden=!1},g=()=>{B(),c("cookie_preference",encodeURIComponent(JSON.stringify(y)),365);var e=document.getElementsByClassName("js-cookie-banner")[0].getAttribute("data-cookie-current-version");c("cookie_preference_set",e,365),k(),document.querySelector(t.confirmationMessageReject).hidden=!1};function h(e,n=""){let i=new Date(0);document.cookie=e+"="+n+";expires="+i+";",document.cookie=e+"="+n+";expires="+i+`;domain=${m};path=/`}var S=()=>{let e=JSON.parse(f("cookie_preference"));return Object.keys(e).filter(n=>e[n]===!0)};function b(e){let n=S();if(a.hasOwnProperty(e))return n.includes(a[e]);let i=Object.keys(a).find(o=>o.endsWith("*")&&e.startsWith(o.slice(0,-1)));if(i){let o=a[i];return n.includes(o)}return!1}function B(){document.cookie.split(";").map(e=>e.trim()).filter(Boolean).forEach(e=>{let n=e.split("=")[0];b(n)||h(n)})}function _(){let e=document.querySelector(t.cookieBanner);e&&(e.hidden=!0)}function k(){let e=document.querySelector(t.cookieBanner),n=document.querySelector(t.confirmationMessageContainer);document.querySelector(t.cookieSelectionContainer).hidden=!0,n.hidden=!1,n.focus(),e.classList.add("cookie-banner--no-decoration")}function j(){if(f("cookie_preference_set"))_();else{let e=document.querySelector(t.cookieBanner);e.hidden=!1,e.removeAttribute("aria-hidden"),c("cookie_preference",encodeURIComponent(JSON.stringify(p)),365)}}function E(){document.querySelector(t.acceptBtn).addEventListener("click",()=>{C(),u()}),document.querySelector(t.rejectBtn).addEventListener("click",()=>{g()}),document.querySelector(t.hideBannerBtn).addEventListener("click",()=>{_()})}function l(){document.querySelector(t.cookieBanner)&&(j(),E())}function d(){let e=document.getElementById("js-enabled-cookie-form");e&&(e.hidden=!1,e.removeAttribute("aria-hidden"))}l();d();})();
+(()=>{var m={essential:!0,analytics:!0,video_players:!1},f={essential:!0,analytics:!1,video_players:!1};var r={_smart_meter_tool_session:"essential",_demo_session:"essential",_content_platform_forms_session:"essential",_energy_apps_session:"essential",cookie_preference:"essential",cookie_preference_set:"essential",_dd_s:"essential",_forms_key:"essential",_grecaptcha:"essential",activeChat:"essential","amazon-connect-*":"essential","amazon-connect-session-*":"essential","cobrowse-storage_expiration-22834971":"essential",CustomizationObject:"essential",cwr_s:"essential",cwr_u:"essential","lpLastVisit-*":"essential",lpPmCalleeDfs:"essential","LPSID-*":"essential",lpTabId:"essential",LPVID:"essential",persistedChatSession:"essential","X-Source":"essential",_ga:"analytics","_ga_*":"analytics",ar_debug:"analytics",ethnio_displayed:"analytics",geo:"analytics","mf_*":"analytics",mf_initialDomQueue:"analytics",mf_transmitQueue:"analytics",mf_user:"analytics",mouseflow:"analytics","ca_ab_*":"analytics",__TAG_ASSISTANT:"analytics"};var _=document.location.hostname==="localhost"?"localhost":"citizensadvice.org.uk",g=()=>{let e=JSON.parse(l("cookie_preference"));return Object.keys(e).filter(n=>e[n]===!0)};function h(e){let n=g();if(Object.prototype.hasOwnProperty.call(r,e))return n.includes(r[e]);let i=Object.keys(r).find(o=>o.endsWith("*")&&e.startsWith(o.slice(0,-1)));if(i){let o=r[i];return n.includes(o)}return!1}function S(e,n=""){let i=new Date(0);document.cookie=e+"="+n+";expires="+i+";",document.cookie=e+"="+n+";expires="+i+`;domain=${_};path=/`}function k(){document.cookie.split(";").map(e=>e.trim()).filter(Boolean).forEach(e=>{let n=e.split("=")[0];h(n)||S(n)})}function c(e,n,i){let o=new Date;o.setTime(o.getTime()+i*24*60*60*1e3);let a="expires="+o.toUTCString();document.cookie=e+"="+n+";"+a+`;path=/;domain=${_}`}function l(e){let n=`${e}=`,i=document.cookie.split(";");for(let o=0,a=i.length;o<a;o++){let s=i[o];for(;s.charAt(0)===" ";)s=s.substring(1,s.length);if(s.indexOf(n)===0)return decodeURIComponent(s.substring(n.length))}return null}function p(){window.dataLayer.push({event:"acceptAllCookies"})}var t={cookieBanner:".js-cookie-banner",acceptBtn:"#js-cookie-banner__button-accept",rejectBtn:"#js-cookie-banner__button-reject",cookieSelectionContainer:".js-cookie-banner__selection-container",confirmationMessageContainer:".js-cookie-banner__confirmation-message",confirmationMessageAccept:".js-cookie-banner__confirmation-accept",confirmationMessageReject:".js-cookie-banner__confirmation-reject",hideBannerBtn:"#js-cookie-banner__button-hide"};function E(){if(l("cookie_preference_set"))C();else{let e=document.querySelector(t.cookieBanner);e.hidden=!1,e.removeAttribute("aria-hidden"),c("cookie_preference",encodeURIComponent(JSON.stringify(m)),365)}}var b=()=>{c("cookie_preference",encodeURIComponent(JSON.stringify({essential:!0,analytics:!0,video_players:!0})),365);var e=document.getElementsByClassName("js-cookie-banner")[0].getAttribute("data-cookie-current-version");c("cookie_preference_set",e,365),y(),document.querySelector(t.confirmationMessageAccept).hidden=!1},O=()=>{k(),c("cookie_preference",encodeURIComponent(JSON.stringify(f)),365);var e=document.getElementsByClassName("js-cookie-banner")[0].getAttribute("data-cookie-current-version");c("cookie_preference_set",e,365),y(),document.querySelector(t.confirmationMessageReject).hidden=!1};function C(){let e=document.querySelector(t.cookieBanner);e&&(e.hidden=!0)}function y(){let e=document.querySelector(t.cookieBanner),n=document.querySelector(t.confirmationMessageContainer);document.querySelector(t.cookieSelectionContainer).hidden=!0,n.hidden=!1,n.focus(),e.classList.add("cookie-banner--no-decoration")}function B(){document.querySelector(t.acceptBtn).addEventListener("click",()=>{b(),p()}),document.querySelector(t.rejectBtn).addEventListener("click",()=>{O()}),document.querySelector(t.hideBannerBtn).addEventListener("click",()=>{C()})}function d(){document.querySelector(t.cookieBanner)&&(E(),B())}function u(){let e=document.getElementById("js-enabled-cookie-form");e&&(e.hidden=!1,e.removeAttribute("aria-hidden"))}d();u();})();
 //# sourceMappingURL=application.js.map
