+++ engine/app/javascript/citizens_advice_cookie_preferences/constants/consent.js
@@ -1,0 +1,11 @@
+export const DEFAULT_COOKIE_CONSENT = {
+  essential: true,
+  analytics: true,
+  video_players: false,
+};
+
+export const ESSENTIAL_COOKIE_CONSENT = {
+  essential: true,
+  analytics: false,
+  video_players: false,
+};
